int numLength(int num) { int length = 0;
{
return (1);
}

while (num)
{
num = num / 10;
length += 1;
}

return (length);
}
/**
 *  *main - prints the first 98 fibonaci sequences
 *  Return: 0
 */

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	unsigned long i, n1, n2, temp;
+	unsigned long long i, n1, n2, temp;
 
 	n1 = 1;
 	n2 = 2;
